refactor: fix review comments
