+++ flow/workflows/cdc_flow.go
@@ -52,18 +52,6 @@ func NewCDCFlowWorkflowState(cfg *protos.FlowConnectionConfigs) *CDCFlowWorkflow
 	}
 }
 
-func GetSideEffect[T any](ctx workflow.Context, f func(workflow.Context) T) T {
-	sideEffect := workflow.SideEffect(ctx, func(ctx workflow.Context) interface{} {
-		return f(ctx)
-	})
-
-	var result T
-	if err := sideEffect.Get(&result); err != nil {
-		panic(err)
-	}
-	return result
-}
-
 func GetUUID(ctx workflow.Context) string {
 	return GetSideEffect(ctx, func(_ workflow.Context) string {
 		return uuid.New().String()
@@ -373,11 +361,11 @@ func CDCFlowWorkflow(
 
 	originalRunID := workflow.GetInfo(ctx).OriginalRunID
 
-	newCtx, err := GetFlowMetadataContext(ctx, cfg.FlowJobName, cfg.SourceName, cfg.DestinationName)
+	var err error
+	ctx, err = GetFlowMetadataContext(ctx, cfg.FlowJobName, cfg.SourceName, cfg.DestinationName)
 	if err != nil {
 		return state, fmt.Errorf("failed to get flow metadata context: %w", err)
 	}
-	ctx = newCtx
 
 	// we cannot skip SetupFlow if SnapshotFlow did not complete in cases where Resync is enabled
 	// because Resync modifies TableMappings before Setup and also before Snapshot
@@ -573,16 +561,4 @@ func CDCFlowWorkflow(
 			return state, workflow.NewContinueAsNewError(ctx, CDCFlowWorkflow, cfg, state)
 		}
 	}
-}
-
-func GetFlowMetadataContext(ctx workflow.Context, flowJobName string, sourceName string, destinationName string) (workflow.Context, error) {
-	metadataCtx := workflow.WithLocalActivityOptions(ctx, workflow.LocalActivityOptions{
-		StartToCloseTimeout: 10 * time.Second,
-	})
-	getMetadataFuture := workflow.ExecuteLocalActivity(metadataCtx, flowable.GetFlowMetadata, flowJobName, sourceName, destinationName)
-	var metadata *protos.FlowContextMetadata
-	if err := getMetadataFuture.Get(metadataCtx, &metadata); err != nil {
-		return nil, err
-	}
-	return workflow.WithValue(ctx, shared.FlowMetadataKey, metadata), nil
 }
