+++ flow/workflows/drop_flow.go
@@ -86,12 +86,12 @@ func DropFlowWorkflow(ctx workflow.Context, input *protos.DropFlowInput) error {
 	ctx = workflow.WithValue(ctx, shared.FlowNameKey, input.FlowJobName)
 	workflow.GetLogger(ctx).Info("performing cleanup for flow",
 		slog.String(string(shared.FlowNameKey), input.FlowJobName))
-	newCtx, err := GetFlowMetadataContext(ctx,
+	var err error
+	ctx, err = GetFlowMetadataContext(ctx,
 		input.FlowJobName, input.FlowConnectionConfigs.SourceName, input.FlowConnectionConfigs.DestinationName)
 	if err != nil {
 		return fmt.Errorf("failed to get flow metadata context: %w", err)
 	}
-	ctx = newCtx
 	if input.FlowConnectionConfigs != nil {
 		if input.DropFlowStats {
 			dropStatsCtx := workflow.WithActivityOptions(ctx, workflow.ActivityOptions{
